+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Master](https://github.com/appwise-labes/AppwiseCore)
 
+### New Features
+
+* Core: Add a new network shortcut for loading a `Decodable` object from JSON.
+
 ## [0.10.1](https://github.com/appwise-labs/AppwiseCore/releases/tag/0.10.1)
 
 ### Bug Fixes